test(e2e-mw-dev): port the kind k8s suite into the real-cluster harness; retire tests/k8s
The kind-based tests/k8s/ suite tested the multi-tenant activation pipeline
against a fake cluster — it could not exercise the layers where this quarter's
bugs actually lived (real Cilium, Crossplane ducklings, cnpg-shard + external
RDS metadata, per-org Lakekeeper). The per-PR mw-dev harness can. Port every
cluster test into harness.sh, implement the remaining TODOs, and remove the
kind suite + its CI job.

harness.sh (now runs kubectl in-cluster via the Job's SA token):
- wire/query: SELECT 1 + 5 concurrent distinct connections
- activation: DuckLake + Iceberg attach + R/W on cnpg AND ext backends
- extension forks: bundled ducklake/httpfs are the PostHog forks, not upstream
- worker pods: labels, securityContext (non-root/uid1000/no-priv-esc),
  Downward-API POD_NAME/NODE_NAME env, no ambient SA-token mount
- resilience: worker-pod kill crash recovery; DuckLake durability across a
  worker restart; concurrent writers (fork conflict-retry)
- isolation: cnpg vs ext see distinct catalogs, cross-tenant read denied
- lifecycle: deprovision -> wait Duckling CR --for=delete -> re-provision the
  SAME org id -> R/W again (the stranded-cnpg-role regression net, now reliable
  because it waits on the CR's finalizer cascade instead of warehouse=deleted)

run.sh: add a `janitor` subcommand; e2e-mw-dev.yml gains a 6h `schedule`
trigger that runs only the janitor (reaps stale duckgres-ci-pr-* namespaces +
their ducklings/cnpg-role/PIA/bindings). NAMESPACE no longer required for janitor.

tests/k8s removal:
- delete the kind Go suite + its testdata + CLAUDE.md (cluster tests ported;
  test-only harness helpers retired with it)
- the RBAC + network-policy static-manifest asserts (the only unit tests over
  real shipped k8s/ config) move to tests/manifests/ and run in `go test ./...`
- remove the k8s-integration-tests job from ci.yml

The supporting k8s/ scripts/manifests + Dockerfiles are kept for now (a later
cleanup PR removes the now-dangling `just test-k8s-integration` recipe).
Deliberately not ported: warm-pool activation + version-reaper (per-PR CP runs
warm-target=0), physical S3-prefix isolation (no list creds from the Job),
Cilium egress probing (needs a stable in-worker exec) — documented in README.
