+++ .github/workflows/ci.yml
@@ -247,88 +247,3 @@ jobs:
 
       - name: Run Kubernetes-tagged control plane package tests
         run: just test-controlplane-k8s
-
-  k8s-integration-tests:
-    needs: unit-tests
-    runs-on: ubuntu-24.04-arm
-    timeout-minutes: 30
-    # id-token: write lets this job request a GitHub OIDC token, which
-    # aws-actions/configure-aws-credentials trades for STS-vended AWS
-    # credentials by assuming github-duckgres-iceberg-ci-testing-role in mw-dev.
-    # The role's trust policy is scoped to repo:PostHog/duckgres:*; its
-    # IAM policy is scoped to the iceberg test buckets only. Provisioned
-    # by PostHog/posthog-cloud-infra#8124.
-    permissions:
-      id-token: write
-      contents: read
-    env:
-      DUCKGRES_KIND_CLUSTER_NAME: duckgres
-      DUCKGRES_KIND_NODE_IMAGE: kindest/node:v1.31.0@sha256:53df588e04085fd41ae12de0c3fe4c72f7013bba32a20e7325357a1ac94ba865
-      # Iceberg integration test (tests/k8s/iceberg_test.go) fails openly
-      # when any of these is unset — see its godoc for the rationale. The
-      # AWS_* credentials are populated by configure-aws-credentials below
-      # via OIDC; the three iceberg-specific values are bucket coordinates
-      # provisioned in mw-dev.
-      DUCKGRES_K8S_ICEBERG_TABLE_BUCKET_ARN: arn:aws:s3tables:us-east-1:${{ secrets.MW_DEV_ACCOUNT_ID }}:bucket/posthog-duckgres-iceberg-test-mw-dev
-      DUCKGRES_K8S_ICEBERG_REGION: us-east-1
-      DUCKGRES_K8S_ICEBERG_DATA_BUCKET: posthog-duckgres-iceberg-test-data-mw-dev
-
-    services:
-      postgres:
-        image: public.ecr.aws/docker/library/postgres:16-alpine
-        env:
-          POSTGRES_USER: postgres
-          POSTGRES_PASSWORD: postgres
-          POSTGRES_DB: testdb
-        ports:
-          - 127.0.0.1:35432:5432
-        options: >-
-          --health-cmd pg_isready
-          --health-interval 10s
-          --health-timeout 5s
-          --health-retries 5
-
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
-      - name: Set up just
-        uses: extractions/setup-just@dd310ad5a97d8e7b41793f8ef055398d51ad4de6 # v2.0.0
-      - name: Set up Go
-        uses: actions/setup-go@f111f3307d8850f501ac008e886eec1fd1932a34 # v5.3.0
-        with:
-          go-version-file: go.mod
-      - name: Install kind
-        env:
-          KIND_VERSION: v0.24.0
-        run: |
-          curl -sSL -o /tmp/kind "https://github.com/kubernetes-sigs/kind/releases/download/${KIND_VERSION}/kind-linux-arm64"
-          chmod +x /tmp/kind
-          sudo mv /tmp/kind /usr/local/bin/kind
-          kind --version
-      - name: Clear Docker Hub credentials for kind pulls
-        run: |
-          docker logout registry-1.docker.io || true
-          docker logout docker.io || true
-          docker logout https://index.docker.io/v1/ || true
-      - name: Pre-pull kind node image
-        run: |
-          for attempt in 1 2 3; do
-            if docker pull "${DUCKGRES_KIND_NODE_IMAGE}"; then
-              exit 0
-            fi
-            sleep $((attempt * 5))
-          done
-          exit 1
-      - name: Configure AWS credentials via OIDC
-        # Trades the GitHub-issued OIDC token for STS credentials by
-        # assuming github-duckgres-iceberg-ci-testing-role in mw-dev. Exposes
-        # AWS_ACCESS_KEY_ID / AWS_SECRET_ACCESS_KEY / AWS_SESSION_TOKEN
-        # in the job env, which iceberg_test.go reads via os.Getenv.
-        # Pinned to the same commit cloud-infra workflows use.
-        uses: aws-actions/configure-aws-credentials@61815dcd50bd041e203e49132bacad1fd04d2708 # v4.0.2
-        with:
-          aws-region: us-east-1
-          role-to-assume: arn:aws:iam::${{ secrets.MW_DEV_ACCOUNT_ID }}:role/github-duckgres-iceberg-ci-testing-role
-          role-duration-seconds: 3600
-      - name: Run Kubernetes integration tests
-        run: just test-k8s-integration
