+++ tests/e2e-mw-dev/run.sh
@@ -14,7 +14,9 @@ set -euo pipefail
 
 HERE="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 CTX="${KUBE_CONTEXT:?}"
-NS="${NAMESPACE:?}"
+# NAMESPACE is required by deploy|test|diagnostics|teardown but NOT by
+# e2e-cleanup (which discovers stale namespaces itself). Don't require it here.
+NS="${NAMESPACE:-}"
 KUBECTL=(kubectl --context "$CTX")
 EKS_CLUSTER_NAME="${EKS_CLUSTER_NAME:-posthog-mw-dev}"
 AWS_REGION="${AWS_REGION:-us-east-1}"
@@ -202,7 +204,7 @@ cmd_diagnostics() {
 cmd_teardown() {
   # Deprovision the ci-pr ducklings FIRST so shared-infra resources (S3 bucket,
   # cnpg role+db, lakekeeper CR/secret/SA) are cleaned up by the control plane
-  # before we delete it. Best-effort — the namespace delete + janitor are the
+  # before we delete it. Best-effort — the namespace delete + e2e-cleanup are the
   # backstop. Uses the CP admin API via a short-lived port-forward.
   if "${KUBECTL[@]}" -n "$NS" get deploy/duckgres-control-plane >/dev/null 2>&1; then
     secret="$(cat "$internal_secret_file" 2>/dev/null || true)"
@@ -254,10 +256,49 @@ cmd_teardown() {
   "${KUBECTL[@]}" delete namespace "$NS" --ignore-not-found --wait=false
 }
 
-case "${1:?usage: run.sh deploy|test|diagnostics|teardown}" in
-  deploy) cmd_deploy ;;
-  test) cmd_test ;;
-  diagnostics) cmd_diagnostics ;;
-  teardown) cmd_teardown ;;
+# Sweep stale per-PR namespaces left behind by runs that died hard (cancelled
+# mid-flight, runner OOM, etc.) before their always() teardown could fire.
+# Backstop, not the primary cleanup. Discovers namespaces by the
+# managed-by=e2e-mw-dev label and deletes any older than E2E_CLEANUP_MAX_AGE_HOURS
+# (default 6h — a real run finishes in <40m, so anything older is orphaned).
+# For each: delete its ducklings directly (drives the Crossplane teardown
+# without needing the now-gone per-run CP), drop the cnpg role+db, drop the Pod
+# Identity association, and sweep the ci-pr-labelled cross-ns bindings + the ns.
+# Named e2e-cleanup (not "janitor") to avoid colliding with duckgres's own
+# control-plane janitor. NAMESPACE is not required for this path.
+cmd_e2e_cleanup() {
+  local max_age_h now ns created age pr
+  max_age_h="${E2E_CLEANUP_MAX_AGE_HOURS:-6}"
+  now="$(date +%s)"
+  "${KUBECTL[@]}" get ns -l app.kubernetes.io/managed-by=e2e-mw-dev \
+    -o jsonpath='{range .items[*]}{.metadata.name}{" "}{.metadata.creationTimestamp}{"\n"}{end}' 2>/dev/null \
+  | while read -r ns created; do
+      [ -n "$ns" ] || continue
+      age=$(( (now - $(date -d "$created" +%s)) / 3600 ))
+      if [ "$age" -lt "$max_age_h" ]; then
+        echo "e2e-cleanup: keep $ns (age ${age}h < ${max_age_h}h)"; continue
+      fi
+      pr="${ns#duckgres-ci-pr-}"
+      echo "e2e-cleanup: reaping $ns (age ${age}h, PR $pr)"
+      for org in "ci-pr-${pr}-cnpg" "ci-pr-${pr}-ext"; do
+        "${KUBECTL[@]}" -n ducklings delete "duckling/$org" --ignore-not-found --wait=false 2>/dev/null || true
+      done
+      for org in "ci-pr-${pr}-cnpg" "ci-pr-${pr}-ext"; do
+        "${KUBECTL[@]}" -n ducklings wait --for=delete "duckling/$org" --timeout=300s 2>/dev/null || true
+      done
+      drop_cnpg_role "ci-pr-${pr}-cnpg"
+      NS="$ns" delete_pod_identity
+      "${KUBECTL[@]}" delete clusterrolebinding -l "duckgres.posthog.com/ci-pr=${pr}" --ignore-not-found
+      "${KUBECTL[@]}" -n lakekeeper delete rolebinding -l "duckgres.posthog.com/ci-pr=${pr}" --ignore-not-found
+      "${KUBECTL[@]}" delete namespace "$ns" --ignore-not-found --wait=false
+    done
+}
+
+case "${1:?usage: run.sh deploy|test|diagnostics|teardown|e2e-cleanup}" in
+  deploy) : "${NAMESPACE:?}"; cmd_deploy ;;
+  test) : "${NAMESPACE:?}"; cmd_test ;;
+  diagnostics) : "${NAMESPACE:?}"; cmd_diagnostics ;;
+  teardown) : "${NAMESPACE:?}"; cmd_teardown ;;
+  e2e-cleanup) cmd_e2e_cleanup ;;
   *) echo "unknown: $1" >&2; exit 2 ;;
 esac
